refactor: extract duplicate test helpers into IntegrationTestBase
