+++ src/test/java/com/iyte_yazilim/proje_pazari/IntegrationTestBase.java
@@ -2,8 +2,12 @@ package com.iyte_yazilim.proje_pazari;
 
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.when;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.iyte_yazilim.proje_pazari.application.commands.loginUser.LoginUserCommand;
+import com.iyte_yazilim.proje_pazari.application.commands.registerUser.RegisterUserCommand;
 import com.iyte_yazilim.proje_pazari.domain.interfaces.TokenBlacklistService;
 import com.iyte_yazilim.proje_pazari.infrastructure.persistence.EmailVerificationRepository;
 import com.iyte_yazilim.proje_pazari.infrastructure.persistence.ProjectApplicationRepository;
@@ -14,16 +18,19 @@ import com.iyte_yazilim.proje_pazari.infrastructure.security.config.RateLimitCon
 import io.github.bucket4j.Bandwidth;
 import io.github.bucket4j.Bucket;
 import java.time.Duration;
+import java.time.LocalDateTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.webmvc.test.autoconfigure.AutoConfigureMockMvc;
 import org.springframework.data.redis.core.StringRedisTemplate;
+import org.springframework.http.MediaType;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 @ActiveProfiles("test")
@@ -43,13 +50,13 @@ public abstract class IntegrationTestBase {
 
     @Autowired private RefreshTokenRepository refreshTokenRepository;
 
-    @Autowired private EmailVerificationRepository emailVerificationRepository;
+    @Autowired protected EmailVerificationRepository emailVerificationRepository;
 
     @Autowired private ProjectApplicationRepository projectApplicationRepository;
 
     @Autowired private ProjectRepository projectRepository;
 
-    @Autowired private UserRepository userRepository;
+    @Autowired protected UserRepository userRepository;
 
     @BeforeEach
     void setUp() {
@@ -69,5 +76,45 @@ public abstract class IntegrationTestBase {
                                         .build())
                         .build();
         when(rateLimitConfig.resolveBucket(anyString())).thenReturn(permissiveBucket);
+    }
+
+    // ── Shared Helper Methods ────────────────────────────────────────────
+
+    protected void registerAndVerifyUser(
+            String email, String password, String firstName, String lastName) throws Exception {
+        var command = new RegisterUserCommand(email, password, firstName, lastName);
+        mockMvc.perform(
+                        post("/api/v1/auth/register")
+                                .contentType(MediaType.APPLICATION_JSON)
+                                .content(objectMapper.writeValueAsString(command)))
+                .andExpect(status().isCreated());
+
+        var user = userRepository.findByEmail(email).orElseThrow();
+        var verification =
+                emailVerificationRepository
+                        .findTopByUserIdOrderByCreatedAtDesc(user.getId())
+                        .orElseThrow();
+        verification.setVerifiedAt(LocalDateTime.now());
+        emailVerificationRepository.save(verification);
+    }
+
+    protected String loginAndGetToken(String email, String password) throws Exception {
+        var command = new LoginUserCommand(email, password);
+        MvcResult result =
+                mockMvc.perform(
+                                post("/api/v1/auth/login")
+                                        .contentType(MediaType.APPLICATION_JSON)
+                                        .content(objectMapper.writeValueAsString(command)))
+                        .andExpect(status().isOk())
+                        .andReturn();
+
+        var jsonNode = objectMapper.readTree(result.getResponse().getContentAsString());
+        return jsonNode.get("data").get("accessToken").asText();
+    }
+
+    protected String createVerifiedUserAndGetToken(
+            String email, String password, String firstName, String lastName) throws Exception {
+        registerAndVerifyUser(email, password, firstName, lastName);
+        return loginAndGetToken(email, password);
     }
 }
