+++ src/test/java/com/iyte_yazilim/proje_pazari/presentation/controllers/UserControllerIntegrationTest.java
@@ -6,12 +6,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.iyte_yazilim.proje_pazari.IntegrationTestBase;
 import com.iyte_yazilim.proje_pazari.application.commands.loginUser.LoginUserCommand;
-import com.iyte_yazilim.proje_pazari.application.commands.registerUser.RegisterUserCommand;
-import com.iyte_yazilim.proje_pazari.infrastructure.persistence.EmailVerificationRepository;
-import com.iyte_yazilim.proje_pazari.infrastructure.persistence.UserRepository;
 import com.iyte_yazilim.proje_pazari.infrastructure.persistence.models.UserEntity;
 import com.iyte_yazilim.proje_pazari.presentation.security.JwtUtil;
-import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
@@ -21,7 +17,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.test.web.servlet.MvcResult;
 
 class UserControllerIntegrationTest extends IntegrationTestBase {
 
@@ -31,56 +26,15 @@ class UserControllerIntegrationTest extends IntegrationTestBase {
     private static final String VALID_FIRST_NAME = "John";
     private static final String VALID_LAST_NAME = "Doe";
 
-    @Autowired private UserRepository userRepository;
-
-    @Autowired private EmailVerificationRepository emailVerificationRepository;
-
     @Autowired private PasswordEncoder passwordEncoder;
 
     @Autowired private JwtUtil jwtUtil;
 
     // ── Helper Methods ──────────────────────────────────────────────────
 
-    private void registerUser(String email, String password, String firstName, String lastName)
-            throws Exception {
-        var command = new RegisterUserCommand(email, password, firstName, lastName);
-        mockMvc.perform(
-                        post("/api/v1/auth/register")
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(command)))
-                .andExpect(status().isCreated());
-    }
-
-    private void registerAndVerifyUser(
-            String email, String password, String firstName, String lastName) throws Exception {
-        registerUser(email, password, firstName, lastName);
-
-        var user = userRepository.findByEmail(email).orElseThrow();
-        var verification =
-                emailVerificationRepository
-                        .findTopByUserIdOrderByCreatedAtDesc(user.getId())
-                        .orElseThrow();
-        verification.setVerifiedAt(LocalDateTime.now());
-        emailVerificationRepository.save(verification);
-    }
-
-    private String loginAndGetToken(String email, String password) throws Exception {
-        var command = new LoginUserCommand(email, password);
-        MvcResult result =
-                mockMvc.perform(
-                                post("/api/v1/auth/login")
-                                        .contentType(MediaType.APPLICATION_JSON)
-                                        .content(objectMapper.writeValueAsString(command)))
-                        .andExpect(status().isOk())
-                        .andReturn();
-
-        var jsonNode = objectMapper.readTree(result.getResponse().getContentAsString());
-        return jsonNode.get("data").get("accessToken").asText();
-    }
-
     private String createVerifiedUserAndGetToken() throws Exception {
-        registerAndVerifyUser(VALID_EMAIL, VALID_PASSWORD, VALID_FIRST_NAME, VALID_LAST_NAME);
-        return loginAndGetToken(VALID_EMAIL, VALID_PASSWORD);
+        return createVerifiedUserAndGetToken(
+                VALID_EMAIL, VALID_PASSWORD, VALID_FIRST_NAME, VALID_LAST_NAME);
     }
 
     private String getUserId(String email) {
